chore: move attribute types to new dir
