+++ packages/amplify_auth_plugin_interface/lib/src/Attribute/AuthNextUpdateAttributeStep.dart
@@ -14,7 +14,6 @@
  */
 
 import 'package:amplify_auth_plugin_interface/amplify_auth_plugin_interface.dart';
-import 'package:flutter/foundation.dart';
 
 /// This wraps all the details around the next step in the user attribute update process.
 ///
