+++ packages/amplify_auth_plugin_interface/lib/src/Attribute/ConfirmUserAttributeResult.dart
