+++ packages/amplify_auth_plugin_interface/lib/src/Attribute/ResendUserAttributeConfirmationCodeResult.dart
@@ -13,7 +13,6 @@
  * permissions and limitations under the License.
  */
 
-import 'package:flutter/foundation.dart';
 import '../types/AuthCodeDeliveryDetails.dart';
 
 /// Wraps the result of a resend user confirmation code operation.
