+++ packages/amplify_auth_plugin_interface/lib/src/Attribute/UpdateUserAttributeResult.dart
@@ -14,7 +14,6 @@
  */
 
 import 'package:amplify_auth_plugin_interface/amplify_auth_plugin_interface.dart';
-import 'package:flutter/foundation.dart';
 
 /// Wraps the result of an update user attribute operation.
 class UpdateUserAttributeResult {
