+++ packages/amplify_auth_plugin_interface/lib/src/Attribute/UpdateUserAttributesRequest.dart
@@ -13,7 +13,6 @@
  * permissions and limitations under the License.
  */
 
-import 'package:flutter/foundation.dart';
 import 'AuthUserAttribute.dart';
 
 /// Encapsulates parameters for a update user attributes operation
